feat: Add Orders page and fix payment status updates
- Add new OrdersPage component with order listing and details
- Fix backend payment status field mapping (total_amount vs total)
- Add environment variable support for frontend URL in PayOS
- Improve webhook logging and error handling
- Add manual payment status update endpoint for testing
- Fix query parameter type conversion in orders service
- Update PaymentSuccessPage to navigate to specific order
- Add deployment guide with environment variable documentation

+++ backend/src/middlewares/orders.middlewares.ts
@@ -78,8 +78,8 @@ export const cancelOrderValidator = [
 // Validation for getting orders query
 export const getOrdersQueryValidator = [
   query('status').optional().isIn(Object.values(OrderStatus)).withMessage(ORDERS_MESSAGES.INVALID_ORDER_STATUS),
-  query('page').optional().isInt({ min: 1 }).withMessage(ORDERS_MESSAGES.INVALID_PAGE),
-  query('limit').optional().isInt({ min: 1, max: 100 }).withMessage(ORDERS_MESSAGES.INVALID_LIMIT),
+  query('page').optional().isInt({ min: 1 }).toInt().withMessage(ORDERS_MESSAGES.INVALID_PAGE),
+  query('limit').optional().isInt({ min: 1, max: 100 }).toInt().withMessage(ORDERS_MESSAGES.INVALID_LIMIT),
   query('sort')
     .optional()
     .isIn(['created_at', 'updated_at', 'total_amount'])
